+++ lib/mongoid/uuid/rspec/shared/examples/concerns.rb
@@ -1,13 +1,13 @@
 RSpec.shared_examples 'Mongoid::UUID::Concerns' do
-  describe 'fields' do
+  describe 'fields', type: :feature do
     it { is_expected.to have_field(:uuid).of_type BSON::Binary }
   end
 
-  describe 'validations' do
+  describe 'validations', type: :feature do
     it { is_expected.to validate_uniqueness_of :uuid }
   end
 
-  describe 'indexes' do
+  describe 'indexes', type: :feature do
     it { is_expected.to have_index_for(uuid: 1).with_options name: 'uuid_index' }
   end
 end